+++ spec/spec_helper.rb
@@ -3,6 +3,7 @@ require_relative "support/rails_helper"
 require_relative "support/public_key_cached_resolver_stub"
 require_relative "support/public_key_resolver_stub"
 require_relative "support/controllable_public_key_resolver_stub"
+require_relative "support/http_client_stub"
 require "timecop"
 require "byebug"