+++ spec/support/http_client_stub.rb
@@ -1,0 +1,18 @@
+module KeycloakApiRails
+  class HTTPClientStub
+    def initialize
+      @realms = []
+      @mutex  = Mutex.new
+    end
+
+    def realms
+      @mutex.synchronize { @realms.dup }
+    end
+
+    def get(realm_id, _path)
+      @mutex.synchronize { @realms << realm_id }
+
+      { "keys" => [] }
+    end
+  end
+end